+++ projects/ui-kit/navigation/ui-bottom-tab-bar/src/public-api.ts
@@ -1,0 +1,5 @@
+/*
+ * Public API Surface of ui-kit/ui-bottom-tab-bar
+ */
+
+export * from './lib/ui-bottom-tab-bar';